--- journeys/et/dockerSetup.ts
@@ -1,9 +1,0 @@
-import { ensureUp } from 'app/et/docker'
-import { Journey } from 'types/journey'
-
-export default {
-  disabled: true,
-  group: 'et-docker',
-  text: 'Load new images & spin up docker containers for ET CCD',
-  fn: ensureUp
-} as Journey
